Criação do Controller

+++ Repositories/UserRepository.cs
@@ -1,5 +1,6 @@
 ﻿using API.Data;
 using API.Entities;
+using Microsoft.EntityFrameworkCore;
 
 namespace API.Repositories
 {
@@ -10,6 +11,17 @@ namespace API.Repositories
         public UserRepository(UserDataContext context)
         {
             _context = context;
+        }        
+
+        public async Task<List<UserEntity>> GetUsersAsync()
+        {
+            return await _context.Users.ToListAsync();
+
+        }
+
+        public async Task<UserEntity> GetByIdEntity(int id)
+        {
+            return await _context.Users.FindAsync(id);
         }
 
         public async Task<UserEntity> CreateEntity(UserEntity userEntity)
@@ -20,14 +32,16 @@ namespace API.Repositories
             return userEntity;
         }
 
-        public async Task<UserEntity> GetByIdEntity(int id)
-        {
-            return await _context.Users.FindAsync(id);
-        }
-
         public async Task<UserEntity> UpdateEntity(UserEntity userEntity)
         {
+            var existingUser = await _context.Users.FindAsync(userEntity.Id);
+            if (existingUser == null)
+            {
+                return null;
+            }
+
             _context.Users.Update(userEntity);
+            
             await _context.SaveChangesAsync();
 
             return userEntity;
@@ -35,6 +49,12 @@ namespace API.Repositories
 
         public async Task<UserEntity> DeleteEntity(UserEntity userEntity)
         {
+            var existingUser = await _context.Users.FindAsync(userEntity.Id);
+            if (existingUser == null)
+            {
+                return null;
+            }
+
             _context.Users.Remove(userEntity);
             await _context.SaveChangesAsync();
